DOCS finiti + cast in `pandos_utils.h`

+++ phase2/include/pandos_utils.h
@@ -39,9 +39,9 @@ bool V(int *);
 
 // Macros for choosing the correct device, used in DOIO
 // ((address - startaddress) / register size) + device starting index
-#define getTypeDevice(cmdAddr) (((int)cmdAddr - 0x10000054) / 0x80 + 3)
+#define getTypeDevice(cmdAddr) (int)(((int)cmdAddr - 0x10000054) / 0x80 + 3)
 //((*cmdAddr - startaddress) - ((type - 3) * register size)) / device n size;
-#define getNumDevice(cmdAddr) ((((int)cmdAddr - 0x10000054) - ((getTypeDevice(cmdAddr) - 3) * 0x80)) / 0x10)
+#define getNumDevice(cmdAddr) (int)((((int)cmdAddr - 0x10000054) - ((getTypeDevice(cmdAddr) - 3) * 0x80)) / 0x10)
 
 // Macro for the timeslice in the format required by setTIMER (aka nubers of ticks)
 #define TIMESLICE_TICKS (TIMESLICE * (*((cpu_t*) TIMESCALEADDR)))
